chore: asserts, diagnostics and transactions moved into client
