+++ amman-client/package.json
@@ -43,8 +43,11 @@
     "registry": "https://registry.npmjs.org"
   },
   "dependencies": {
+    "@metaplex-foundation/cusper": "^0.0.2",
+    "@solana/spl-token-registry": "^0.2.2405",
     "@solana/web3.js": "^1.35.0",
     "debug": "^4.3.3",
+    "deep-diff": "^1.0.2",
     "socket.io-client": "^4.4.1"
   },
   "devDependencies": {
