+++ amman-client/src/consts.ts
