+++ amman-client/src/diagnostics/api.ts
@@ -1,2 +1,1 @@
 export * from './address-labels'
-export * from './token-balances'
