+++ amman-client/src/diagnostics/token-registry.ts
