+++ amman-client/src/transactions/consts.ts
@@ -1,7 +1,5 @@
 import { ConfirmOptions } from '@solana/web3.js'
 
-export * from './types'
-
 /**
  * Default options for sending and confirming a transaction
  * @category transactions
