+++ amman/package.json
@@ -48,19 +48,17 @@
     "registry": "https://registry.npmjs.org"
   },
   "dependencies": {
-    "@metaplex-foundation/cusper": "^0.0.2",
+    "@metaplex-foundation/amman-client": "^0.0.0",
     "@metaplex-foundation/js-next": "^0.9.0",
     "@solana/spl-token": "^0.2.0",
-    "@solana/spl-token-registry": "^0.2.2405",
     "@solana/web3.js": "^1.35.0",
-    "@metaplex-foundation/amman-client": "^0.0.0",
     "ansi-colors": "^4.1.1",
     "bn.js": "^5.2.0",
     "buffer-hexdump": "^1.0.0",
     "date-fns": "^2.28.0",
     "debug": "^4.3.3",
     "deep-diff": "^1.0.2",
-    "diff": "^5.0.0",
+    "diff": "^5.1.0",
     "numeral": "^2.0.6",
     "port-pid": "^0.0.7",
     "socket.io": "^4.4.1",
@@ -74,6 +72,7 @@
     "@types/bn.js": "^5.1.0",
     "@types/debug": "^4.1.7",
     "@types/deep-diff": "^1.0.1",
+    "@types/diff": "^5.0.2",
     "@types/node": "^16.11.12",
     "@types/numeral": "^2.0.2",
     "@types/split2": "^3.2.1",
