+++ amman/src/assets/local-accounts.ts
@@ -1,9 +1,10 @@
 import { PublicKey } from '@solana/web3.js'
 import { Account } from '../validator/types'
 import { spawnSync } from 'child_process'
-import { isValidPublicKeyAddress, logError, logInfo } from '../utils'
+import { logError, logInfo } from '../utils'
 import { canAccess, ensureDirSync } from '../utils/fs'
 import path from 'path'
+import { isValidPublicKeyAddress } from '@metaplex-foundation/amman-client'
 
 export async function getExecutableAddress(programId: string): Promise<string> {
   const programPubkey = new PublicKey(programId)
