+++ amman/src/validator/init-validator.ts
@@ -1,12 +1,9 @@
+import { logError, logInfo, logTrace, sleep, tmpLedgerDir } from '../utils'
+import { killRunningServer, resolveServerAddress } from '../utils/http'
 import {
   LOCALHOST,
-  logError,
-  logInfo,
-  logTrace,
-  sleep,
-  tmpLedgerDir,
-} from '../utils'
-import { killRunningServer, resolveServerAddress } from '../utils/http'
+  AMMAN_STORAGE_PORT,
+} from '@metaplex-foundation/amman-client'
 
 import http from 'http'
 import { execSync as exec, spawn } from 'child_process'
@@ -25,7 +22,6 @@ import { canAccessSync } from '../utils/fs'
 import { processAccounts } from './process-accounts'
 import { mapPersistedAccountInfos } from '../assets'
 import { processSnapshot } from './process-snapshot'
-import { AMMAN_STORAGE_PORT } from '@metaplex-foundation/amman-client'
 
 /**
  * @private
