+++ package.json
@@ -1,8 +1,8 @@
 {
   "scripts": {
     "build": "yarn build:client && yarn build:amman",
-    "build:client": "tsc -p amman-client/tsconfig.json",
-    "build:amman": "tsc -p amman/tsconfig.json",
+    "build:client": "rimraf amman-client/dist && tsc -p amman-client/tsconfig.json",
+    "build:amman": "rimraf amman/dist && tsc -p amman/tsconfig.json",
     "doc": "(cd ./amman && yarn doc);",
     "doc:update": "./sh/update-docs"
   },
@@ -18,6 +18,7 @@
   },
   "private": true,
   "devDependencies": {
+    "rimraf": "^3.0.2",
     "typescript": "^4.7.2"
   }
 }
